portal2/backend/requirements.txt
alabaster==0.7.13
apiclient==1.0.4
asgiref==3.6.0
Babel==2.12.1
black==23.3.0
build==0.10.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-filter==22.1
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-super-inlines==0.2.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
docutils==0.20.1
drf-yasg==1.21.5
et-xmlfile==1.1.0
filelock==3.12.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
identify==2.5.23
idna==3.4
imagesize==1.4.1
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
livereload @ https://github.com/lepture/python-livereload/archive/master.zip
Markdown==3.4.1
MarkupSafe==2.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.3
oauth2client==4.1.3
openpyxl==3.1.2
packaging==23.0
pandas==2.0.1
pathspec==0.11.1
phonenumbers==8.13.7
Pillow==9.4.0
platformdirs==3.5.0
pre-commit==3.2.2
protobuf==4.22.3
# psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pyproject-hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-monkey-business==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
snowballstemmer==2.2.0
sphinx==7.0.1
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
tomli==2.0.1
tornado==6.3.2
transliterate==1.10.2
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.23.0
